Stav k 18.07.2017

+++ backend/obalky/lib/Eshop/Kramerius.pm
@@ -1,4 +1,4 @@
-package Eshop::Kramerius_MZK;
+package Eshop::Kramerius;
 use base 'Eshop::Mechanize';
 use DateTime::Format::ISO8601;
 use DateTime;
@@ -18,38 +18,40 @@ use LWP::Simple;
 use Eshop;
 use DB;
 
-my ($req,$url,$resp,$ua,$xml,$res,$numFound,$query_url,$OBJIdentifiers);
+
+my ($req,$url,$resp,$ua,$xml,$res,$numFound,$query_url,$OBJIdentifiers,$baseURL);
 my $start = 0; 
 __PACKAGE__->register(crawl => 1, license => 'licensed', czech => 0 );
 
 
 sub crawl{
-	my($self,$storable,$from,$to,$tmp_dir,$feed_url,$eshop) = @_;
+	my($self,$eshop,$from,$to,$tmp_dir,$feedURL) = @_;
+	$baseURL = $feedURL . '/search/';
 	my @type = ("soundrecording","archive","graphic","sheetmusic","manuscript","monograph","periodicalitem");
 	my @list;
 	foreach (@type){
-		warn "Crawling $_";
-		my @type_list = crawl_type($self,$storable,$from,$to,$tmp_dir,$feed_url,$eshop, $_);
+		$start = 0;
+		my @type_list = crawl_type($self,$eshop,$from,$to,$tmp_dir, $_);
 		push (@list,@type_list) if (@type_list);
 	}
 	return @list;
 }
 
 sub crawl_type{
-	my($self,$storable,$from,$to,$tmp_dir,$feed_url,$eshop,$method) = @_;
+	my($self,$eshop,$from,$to,$tmp_dir,$method) = @_;
 	my (@books,$root_pid);
 	get_response($from,$to,$method);
 	if ($resp->is_success) {
-			$xml = eval { XMLin($resp->content,SuppressEmpty => 1, ForceArray =>['doc']) };
-			$res = $xml->{'result'};
-			$numFound = $res->{'numFound'};
-			my $numberOfQueries = ceil($numFound / 10);
-			warn Dumper('Pocet zaznamov:' . $numFound. '\n URL : ');#.$query_url);
-		     for (my $i=1;$i <= $numberOfQueries ;$i++){
-		     	 foreach my $xml_item (@{$res->{'doc'}}) {
-		     	my ($PIDxml,$PID,$PIDurl,$PIDua,$PIDreq,$PIDresp,$content);
+		$xml = eval { XMLin($resp->content,SuppressEmpty => 1, ForceArray =>['doc']) };
+		$res = $xml->{'result'};
+		$numFound = $res->{'numFound'};
+		my $numberOfQueries = ceil($numFound / 10);
+		warn "Zaznam: $method 		Pocet:  $numFound" if ($ENV{DEBUG});
+		for (my $i=1;$i <= $numberOfQueries ;$i++){
+			foreach my $xml_item (@{$res->{'doc'}}) {
+				my ($PIDxml,$PID,$PIDurl,$PIDua,$PIDreq,$PIDresp,$content);
 				$PID = $xml_item->{'str'}->{'content'};
-				$PIDurl = 'http://kramerius.mzk.cz/search/api/v5.0/item/' . $PID .'/streams/BIBLIO_MODS';
+				$PIDurl = $baseURL . 'api/v5.0/item/' . $PID .'/streams/BIBLIO_MODS';
 				$PIDua = LWP::UserAgent->new;
 				$PIDua->timeout(60);
 				$PIDreq = HTTP::Request->new(GET => $PIDurl);
@@ -67,8 +69,8 @@ sub crawl_type{
 						case /^(monograph|manuscript|sheetmusic|archive|maprecord|graphic|soundrecording)$/ {
 							$title = $PIDxml->{'mods'}->{'titleInfo'}->[0]->{'title'};
 							$year = $PIDxml->{'mods'}->{'originInfo'}->[0]->{'dateIssued'}->[0];
-							$book_url = 'http://kramerius.mzk.cz/search/api/v5.0/item/' . $PID .'/children';
-							$product_url = 'http://kramerius.mzk.cz/search/i.jsp?pid=' . $PID;
+							$book_url = $baseURL . 'api/v5.0/item/' . $PID .'/children';
+							$product_url = $baseURL . 'i.jsp?pid=' . $PID;
 						    $OBJIdentifiers = $PIDxml->{'mods'}->{'identifier'};
 						    $authorslist = $PIDxml->{'mods'}->{'name'};
 							}
@@ -76,14 +78,14 @@ sub crawl_type{
 						case 'periodicalitem' {
 							my ($PERjsonurl,$PERreq,$PERresp,$ROOTurl,$ROOTreq,$ROOTresp,$ROOTxml,$ROOTcontent);
 												
-							$PERjsonurl = 'http://kramerius.mzk.cz/search/api/v5.0/item/'. $PID;
+							$PERjsonurl = $baseURL . 'api/v5.0/item/'. $PID;
 							$PERreq = HTTP::Request->new(GET => $PERjsonurl);
 							$PERreq->header('content-type' => 'application/json');
 						    $PERresp = $PIDua->request($PERreq);
 						    $jsondata= decode_json($PERresp->content);
 						    #ziskani korenoveho periodika	
 						    $root_pid = $jsondata->{'root_pid'};
-					    	$ROOTurl = 'http://kramerius.mzk.cz/search/api/v5.0/item/' . $root_pid .'/streams/BIBLIO_MODS';
+					    	$ROOTurl = $baseURL . 'api/v5.0/item/' . $root_pid .'/streams/BIBLIO_MODS';
 							$ROOTreq = HTTP::Request->new(GET => $ROOTurl);
 							$ROOTreq->header('content-type' => 'application/xml');
 							$ROOTresp = $PIDua->request($ROOTreq);
@@ -91,8 +93,8 @@ sub crawl_type{
 							$ROOTxml = eval { XMLin($ROOTcontent,SuppressEmpty => 1, ForceArray => ['identifier','name','titleInfo','namePart','originInfo','dateIssued'])};
 							$root_title = $ROOTxml->{'mods'}->{'titleInfo'}->[0]->{'title'};
 							$root_year = $ROOTxml->{'mods'}->{'originInfo'}->[0]->{'dateIssued'}->[0];
-							$root_book_url = 'http://kramerius.mzk.cz/search/api/v5.0/item/' . $root_pid .'/children';
-							$root_product_url = 'http://kramerius.mzk.cz/search/i.jsp?pid=' . $root_pid;
+							$root_book_url = $baseURL . 'api/v5.0/item/' . $root_pid .'/children';
+							$root_product_url = $baseURL . 'i.jsp?pid=' . $root_pid;
 							$OBJIdentifiers = $ROOTxml->{'mods'}->{'identifier'};
 							$authorslist = $ROOTxml->{'mods'}->{'name'};	
 							$title = $PIDxml->{'mods'}->{'titleInfo'}->[0]->{'title'} || $jsondata ->{'title'} ;
@@ -104,18 +106,20 @@ sub crawl_type{
 					#ziskani vsech autoru
 					$authors='';
 					foreach my $author (@{$authorslist}){
-							my $tmpauth;
-							if (ref $author->{'namePart'}[0] eq 'HASH' and $author->{'namePart'}[1]->{'content'}){
-								foreach my $namePart (@{$author->{'namePart'}}){
-									$tmpauth = $namePart->{'content'};
-									last if ($namePart->{'type'} && $namePart->{'type'} eq 'given');
-								}
+						my $tmpauth;
+						if (ref $author->{'namePart'}[0] eq 'HASH' and $author->{'namePart'}[1]->{'content'}){
+							foreach my $namePart (@{$author->{'namePart'}}){
+								$tmpauth = $namePart->{'content'};
+								last if ($namePart->{'type'} && $namePart->{'type'} eq 'given');
 							}
-							else {
-								$tmpauth = $author->{'namePart'}[0];
-							}							
-							
-							$authors = $authors . $tmpauth.", "; #if (ref $author->{'namePart'}[0]->{'content'} eq 'HASH');
+						}
+						else {
+							$tmpauth = $author->{'namePart'}[0];
+							#docasna uprava -- pravdepodobne zmena zo strany Krameria
+							#$tmpauth = $author->{'namePart'}[0]->{'content'};
+						}							
+						
+						$authors = $authors . $tmpauth.", "; #if (ref $author->{'namePart'}[0]->{'content'} eq 'HASH');
 					}
 					if ($authors eq ''){
 						$authors = undef;
@@ -123,7 +127,7 @@ sub crawl_type{
 					else {
 						$authors = substr($authors, 0, -2);
 					}
-								
+					
 					my (@eans,@params);
 					my ($ean,$oclc,$nbn,$sysno,$urnnbn);
 	     			foreach my $identifier (@{$OBJIdentifiers}){
@@ -158,24 +162,24 @@ sub crawl_type{
 					}
 					
 					# zaznamy bez parametrov pouzitelnych i v inych knizniciach nebereme
-					next if (!$ean && !$oclc && !$nbn);
+					#next if (!$ean && !$oclc && !$nbn);
 					
-					#kdyz urnnbn je jediny identifikator					
+					#kdyz urnnbn je jediny identifikator
 					$nbn = $sysno || $urnnbn if (!$oclc && !$nbn && !$ean);
 					#Kdyz objekt obsahuje vice EAN
-					if (@eans){
+					if (@eans) {
 						my $ean = shift(@eans);
-						if (@eans > 0){
-							foreach my $param (@eans){
+						if (@eans > 0) {
+							foreach my $param (@eans) {
 								push @params, { 'ean13'=>$param, 'nbn'=>undef, 'oclc'=>undef } if ($param);
 							}
 						}
 					}
-					else{
+					else {
 						push(@eans,undef);
 					}
-					next if (!$ean && !$oclc && !$nbn);
-					if ($method eq 'periodicalitem'){
+					
+					if ($method eq 'periodicalitem') {
 						$rootbibinfo = Obalky::BibInfo->new_from_params({
 							ean => $ean,
 							nbn => $nbn,
@@ -185,10 +189,11 @@ sub crawl_type{
 							authors => $authors,
 							uuid => $root_pid
 						});
-						$rootinfo->{cover_url} = 'http://kramerius.mzk.cz/search/api/v5.0/item/' . $PID .'/thumb';
+						$rootinfo->{cover_url} = $baseURL . 'api/v5.0/item/' . $PID .'/thumb';
 						$rootmedia = Obalky::Media->new_from_info($rootinfo);
-						$book_url = 'http://kramerius.mzk.cz/search/api/v5.0/item/' . $PID .'/children';
-						$product_url = 'http://kramerius.mzk.cz/search/i.jsp?pid=' . $PID;
+						$book_url = $baseURL . 'api/v5.0/item/' . $PID .'/children';
+						$product_url = $baseURL . 'i.jsp?pid=' . $PID;
+
 						$eshop->add_product($rootbibinfo,$rootmedia,$root_product_url,undef,@params);
 						$book_id = DB->resultset('Book')->find_by_bibinfo($rootbibinfo);
 					}
@@ -211,8 +216,9 @@ sub crawl_type{
 					($jdata, $info->{tocpdf_tmpfile}) = downloadtoc($PID,$PIDua,$tmp_dir,$custom_ean) if ($method eq 'periodicalitem' || $method eq 'monograph');
 					$info->{cover_url} = downloadcover($book_url, $jdata);
 					$media = Obalky::Media->new_from_info($info);
-					push(@books, [$bibinfo, $media, $product_url, undef,\@params  ]);
+					warn Dumper($bibinfo) if ($ENV{DEBUG} == 2);
 					warn Dumper($bibinfo);
+					push(@books, [$bibinfo, $media, $product_url, undef, \@params ]);
 				}
 				
 				else {
@@ -222,27 +228,27 @@ sub crawl_type{
 		    last if ($i == $numberOfQueries);
 	        $start += 10;
 	     	get_response($from,$to,$method);
-	     	if ($resp->is_success){
+	     	if ($resp->is_success) {
 	     		$xml = eval { XMLin($resp->content,SuppressEmpty => 1, ForceArray =>['doc']) };
 				$res = $xml->{'result'};	    					
-	     	}
-	     	else {
-	     		print STDERR $resp->status_line, " - chyba pri dalsom dotaze na zaznam knih";
-	     	}
-	     	
-	   }
-	   return @books; 
+			}
+			else {
+				print STDERR $resp->status_line, " - chyba pri dalsom dotaze na zaznam knih";
+			}
+		
+		}
+		return @books; 
 	}
-		else {    
-			print STDERR $resp->status_line, " - chyba pri dotaze na zoznam knih\n";
+	else {
+		print STDERR $resp->status_line, " - chyba pri dotaze na zoznam knih\n";
 	}	
 }
-sub get_response{
+
+sub get_response {
 	my($from,$to,$method) = @_;
 	$method = "map" if ($method eq 'maprecord');
-    #$query_url = "http://kramerius.mzk.cz/search/api/v5.0/search?q=fedora.model:$method%20AND%20modified_date:[".$from."Z%20TO%20".$to.'Z]&fl=PID&wt=xml' . "&start=$start";
-    $query_url = "http://kramerius.mzk.cz/search/api/v5.0/search?q=fedora.model:$method%20AND%20PID:%22uuid:744fc0a0-79c8-11e2-b212-005056827e52%22&fl=PID&wt=xml" . "&start=$start";
-warn $query_url;
+    $query_url = $baseURL . "api/v5.0/search?q=fedora.model:$method%20AND%20modified_date:[".$from."Z%20TO%20".$to."Z]&fl=PID&wt=xml&start=$start";
+	warn "URL: $query_url" if ($ENV{DEBUG} == 2);
 	$ua = LWP::UserAgent->new;
 	$ua->timeout(60);
 	$req = HTTP::Request->new(GET => $query_url);
@@ -257,10 +263,10 @@ sub getEAN{
 	return $parsedIdentifier;	
 }
 
-sub downloadtoc{
+sub downloadtoc {
 	my ($PID,$PIDua,$tmp_dir,$ean) = @_; 
 	my $toc_dir = "$tmp_dir/$PID";
-	my $PIDjsonurl = 'http://kramerius.mzk.cz/search/api/v5.0/item/'. $PID .'/children';
+	my $PIDjsonurl = $baseURL . 'api/v5.0/item/'. $PID .'/children';
 	my $PIDreq = HTTP::Request->new(GET => $PIDjsonurl);
 	$PIDreq->header('content-type' => 'application/json');
     my $PIDresp = $PIDua->request($PIDreq);
@@ -269,28 +275,27 @@ sub downloadtoc{
     my $index = 0;
     my $toc_page;
     my $toc_file;
-	foreach my $pagemodel (@{$jsondata}){
-		if (($pagemodel->{'details'}->{'type'} eq 'TableOfContents')){
-			if ($index == 0){
+	foreach my $pagemodel (@{$jsondata}) {
+		if (($pagemodel->{'details'}->{'type'}) && ($pagemodel->{'details'}->{'type'} eq 'TableOfContents')) {
+			if ($index == 0) {
 				system("rm -rf $toc_dir");
 				mkdir ($toc_dir	);
 			}
 			$index++;
 			$toc_page = $pagemodel->{'pid'};
-			my $tocurl = "http://kramerius.mzk.cz/search/api/v5.0/item/$toc_page/thumb";
+			my $tocurl = $baseURL . "api/v5.0/item/$toc_page/thumb";
 			if(system("wget -q $tocurl -O $toc_dir/$index >/dev/null")) {
 				warn "$tocurl: failed to wget!\n";
 				return ($jsondata, undef);
 			}
 		}
 	}
-	if ($index != 0){
-		$toc_file = "$tmp_dir/$ean.pdf";
+	if ($index != 0) {
+		$toc_file = "$tmp_dir/$PID.pdf";
 		system("convert $toc_dir/* $toc_file");
 		system("rm -rf $toc_dir");
 	}
-	return ($jsondata, $toc_file);
-	
+	return ($jsondata, $toc_file);	
 }
 
 sub downloadcover{
@@ -304,16 +309,15 @@ sub downloadcover{
     	$jsondata = decode_json($PIDresp->content);
 	}
 	return if (!$jsondata);
-	my $coverpage;
-	my $coverurl;
+	my ($coverpage, $coverurl);
 	foreach my $pagemodel (@{$jsondata}){
 		my $page = $pagemodel->{'details'}->{'type'};
 		next if (!$page || (grep /^$page$/i, ("spine","hrbet")));
 		$coverpage = $pagemodel->{'pid'} if (!$coverpage || $page eq 'FrontCover');
 		last if ($page eq 'FrontCover');
 	}
-    $coverurl = "http://kramerius.mzk.cz/search/api/v5.0/item/$coverpage/thumb" if ($coverpage);
-    #$coverurl = "http://kramerius.mzk.cz/search/api/v5.0/item/$coverpage/full" if ($coverpage);
+    #$coverurl = $baseURL . "api/v5.0/item/$coverpage/full" if ($coverpage); - chybaju autorske prava
+    $coverurl = $baseURL . "api/v5.0/item/$coverpage/thumb" if ($coverpage);
 	return $coverurl;
 	
 }
